# Remove N+1 obj creation in flatten_arranged_rels