@Fryguy 2018-05-08
@NickLaMuro Doesn't this end up modifying the caller?  I was trying to avoid that originally.