@NickLaMuro 2018-05-08
@Fryguy No (at least I don't think), because I am pretty sure both `.keys` and `.values` create new Arrays that reference the keys and values respectively, if I am not mistaken.