@Fryguy 2018-05-08
Yes, the .values part does but the iteration here https://github.com/ManageIQ/manageiq/pull/17325/files#diff-e652e9eb772d8d9d7e873e5d891c0c9bR82 doesn't, I don't think.