@NickLaMuro 2018-05-08
@Fryguy okay, I see where it looks like it does that, but in reality, it doesn't pop it from the `kids` hash, but the `Array` that is created when `.values` is called.  For example:

```ruby
irb:001> hash = {:foo => [{:bar => []}, {:baz => [{:qux => []}]}]}
=> {:foo=>[{:bar=>[]}, {:baz=>[{:qux=>[]}]}]}
irb:002> remaining_children = hash.values
=> [[{:bar=>[]}, {:baz=>[{:qux=>[]}]}]]
```

You will notice that the array that is created is a "doubly nested" at the top level.  This means that when `.pop` is called on it it will return:

```ruby
#=> [{:bar=>[]}, {:baz=>[{:qux=>[]}]}]
```

which is the `value` of the `:foo` key.  When `<<` is called on the `kids` of `:bar`, the array will then become:

```ruby
#=> [[]]
```

So we are only mutating the `.values` array that we created within this method, and passing around a bunch of object references otherwise.  Make sense?