@NickLaMuro 2018-05-08
I think another thing to note is that we are doing a `.pop` and iterating directly on the result, instead of doing a `.pop` inside of the iteration which is what is more commonly done (the "iteration" being the `each` in this case, not the surrounding `while`/`until`).