@Fryguy 2018-05-09
This is good :+1:

I think your example Hash there threw me a bit, because it's not accurate.  So, I instrumented and got the following

```
[8] pry(main)> Relationship.flatten_arranged_rels({:foo => {:bar => {}, :baz => {:qux => {}}}})
relationships: 70297681056560
relationships[:foo]: 70297681056600
relationships[:foo][:bar]: 70297681056680
relationships[:foo][:baz]: 70297681056620
relationships[:foo][:baz][:qux]: 70297681056660
remaining_children: [{:bar=>{}, :baz=>{:qux=>{}}}] [70297681056600]
rel: bar 4149468
kids: {} 70297681056680
rel: baz 14284188
kids: {:qux=>{}} 70297681056620
remaining_children: [{}, {:qux=>{}}] [70297681056680, 70297681056620]
rel: qux 14284508
kids: {} 70297681056660
remaining_children: [{}, {}] [70297681056680, 70297681056660]
remaining_children: [{}] [70297681056680]
=> [:foo, :bar, :baz, :qux]
```

Those numbers are object_ids, and you can see that while the "real" values are being used directly as I expected, they are ultimately only being shuffled around, and not modified directly.  As you said, the .pop is only called on the outer array only, which is a copy anyway.