@simaishi 2018-06-21
Added `fine/yes` and `gaprindashvili/yes` as per https://bugzilla.redhat.com/show_bug.cgi?id=1569626