@simaishi 2018-06-21
Fine backport details:

```
$ git log -1
commit a6dc75ee3d0eb686d8c3dab9c37b8a4a4865a6d9
Author: Keenan Brock <keenan@thebrocks.net>
Date:   Fri May 4 10:40:36 2018 -0400

    Merge pull request #17325 from NickLaMuro/flatten_arranged_rels_optimization
    
    Remove N+1 obj creation in flatten_arranged_rels
    (cherry picked from commit 32fadeb91c0e63e34efa4c9321ec52a6bdb176ae)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1593798
```