@simaishi 2018-06-21
Gaprindashvili backport details:

```
$ git log -1
commit 245963eed00817208e34ae869f25f90ddf383216
Author: Keenan Brock <keenan@thebrocks.net>
Date:   Fri May 4 10:40:36 2018 -0400

    Merge pull request #17325 from NickLaMuro/flatten_arranged_rels_optimization
    
    Remove N+1 obj creation in flatten_arranged_rels
    (cherry picked from commit 32fadeb91c0e63e34efa4c9321ec52a6bdb176ae)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1593797
```